+++ thaniya_common/src/thaniya_common/utils/__init__.py
@@ -6,3 +6,6 @@ __version__ = "0.2021.1.18"
 
 
 from .IOContext import IOContext
+from .APIPassword import APIPassword
+
+